+++ src/db/migrations/0000_special_next_avengers.sql
@@ -1,0 +1,30 @@
+CREATE TABLE `collections` (
+	`id` text PRIMARY KEY NOT NULL,
+	`created_at` integer NOT NULL,
+	`updated_at` integer NOT NULL,
+	`deleted_at` integer,
+	`user_id` text,
+	`name` text NOT NULL
+);
+--> statement-breakpoint
+CREATE TABLE `post_collections` (
+	`id` text PRIMARY KEY NOT NULL,
+	`created_at` integer NOT NULL,
+	`updated_at` integer NOT NULL,
+	`deleted_at` integer,
+	`user_id` text,
+	`post_id` text NOT NULL,
+	`collection_id` text NOT NULL,
+	FOREIGN KEY (`post_id`) REFERENCES `posts`(`id`) ON UPDATE no action ON DELETE no action,
+	FOREIGN KEY (`collection_id`) REFERENCES `collections`(`id`) ON UPDATE no action ON DELETE no action
+);
+--> statement-breakpoint
+CREATE TABLE `posts` (
+	`id` text PRIMARY KEY NOT NULL,
+	`created_at` integer NOT NULL,
+	`updated_at` integer NOT NULL,
+	`deleted_at` integer,
+	`user_id` text,
+	`url` text,
+	`note` text
+);